AppServer/requirements.txt
black
Flask
jsonify
pylint
pylint-flask
pyrebase4
pillow
requests